requirements.txt
fastapi
uvicorn
langchain==0.1.14
langchain-experimental
langchain-core
langchain-community
psycopg2-binary
python-dotenv